requirements.txt
tensorflow
fastapi
uvicorn
numpy
pandas
scikit-learn
matplotlib
pillow
pandas
openpyxl